# backend/docs/swagger/api-json.json
{"openapi":"3.0.0","paths":{"/users":{"get":{"operationId":"UsersController_getAllUsers","parameters":[],"responses":{"200":{"description":"List of users"}},"summary":"Get all users","tags":["Users"]},"post":{"operationId":"UsersController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User created"},"400":{"description":"Invalid input"}},"summary":"Create a new user","tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UsersController_getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"User found"},"404":{"description":"User not found"}},"summary":"Get user by ID","tags":["Users"]},"put":{"operationId":"UsersController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/updateUserDto"}}}},"responses":{"200":{"description":"User updated"},"404":{"description":"User not found"}},"summary":"Update a user","tags":["Users"]},"delete":{"operationId":"UsersController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"User deleted"},"404":{"description":"User not found"}},"summary":"Delete a user","tags":["Users"]}},"/":{"get":{"operationId":"AuthController_getHello","parameters":[],"responses":{"200":{"description":"Returns Hello World"}},"summary":"Get Hello World","tags":["Auth"]}},"/recipes":{"get":{"operationId":"RecipesController_getAllRecipes","parameters":[],"responses":{"200":{"description":"List of recipes"}},"summary":"Get all recipes","tags":["Recipes"]},"post":{"operationId":"RecipesController_createRecipe","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRecipeDto"}}}},"responses":{"201":{"description":"Recipe created"},"400":{"description":"Invalid input"}},"summary":"Create a new recipe","tags":["Recipes"]}},"/recipes/{id}":{"get":{"operationId":"RecipesController_getRecipe","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"The recipe"},"404":{"description":"Recipe not found"}},"summary":"Get a recipe by ID","tags":["Recipes"]},"put":{"operationId":"RecipesController_updateRecipe","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRecipeDto"}}}},"responses":{"200":{"description":"Recipe updated"},"400":{"description":"Invalid input"},"404":{"description":"Recipe not found"}},"summary":"Update a recipe","tags":["Recipes"]},"delete":{"operationId":"RecipesController_deleteRecipe","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Recipe deleted"},"404":{"description":"Recipe not found"}},"summary":"Delete a recipe","tags":["Recipes"]}},"/recipes/{id}/public":{"get":{"operationId":"RecipesController_getPublicRecipe","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The public recipe"},"404":{"description":"Recipe not found"}},"summary":"Get a public recipe by its public ID","tags":["Recipes"]}},"/recipes/user/{userId}":{"get":{"operationId":"RecipesController_getRecipesByUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"List of recipes"},"404":{"description":"User not found"}},"summary":"Get recipes by user ID","tags":["Recipes"]}},"/rating":{"get":{"operationId":"RatingController_getAllRatings","parameters":[],"responses":{"200":{"description":"List of ratings"}},"summary":"Get all ratings","tags":["Rating"]},"post":{"operationId":"RatingController_createRating","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRatingDto"}}}},"responses":{"201":{"description":"Rating created"},"400":{"description":"Invalid input"}},"summary":"Create a new rating","tags":["Rating"]}},"/rating/{id}":{"get":{"operationId":"RatingController_getRating","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"The rating with the specified ID"},"404":{"description":"Rating not found"}},"summary":"Get a rating by ID","tags":["Rating"]},"put":{"operationId":"RatingController_updateRating","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRatingDto"}}}},"responses":{"200":{"description":"Rating updated"},"400":{"description":"Invalid input"},"404":{"description":"Rating not found"}},"summary":"Update a rating","tags":["Rating"]},"delete":{"operationId":"RatingController_deleteRating","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Rating deleted"},"404":{"description":"Rating not found"}},"summary":"Delete a rating","tags":["Rating"]}},"/rating/recipe/{recipeId}":{"get":{"operationId":"RatingController_getRatingsByRecipe","parameters":[{"name":"recipeId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"List of ratings for the specified recipe"},"404":{"description":"Recipe not found"}},"summary":"Get ratings for a specific recipe","tags":["Rating"]}}},"info":{"title":"Recipes API","description":"API para recetas y ratings","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"firstName":{"type":"string","description":"Primer nombre del usuario","example":"Pablo"},"lastName":{"type":"string","description":"Apellido del usuario","example":"López"},"email":{"type":"string","description":"Correo electrónico del usuario","example":"pablo.lopez@example.com"},"password":{"type":"string","description":"Contraseña del usuario","example":"password123"}},"required":["firstName","lastName","email","password"]},"updateUserDto":{"type":"object","properties":{"firstName":{"type":"string","description":"Primer nombre del usuario","example":"Pablo"},"lastName":{"type":"string","description":"Apellido del usuario","example":"López"},"email":{"type":"string","description":"Correo electrónico del usuario","example":"pablo.lopez@example.com"},"password":{"type":"string","description":"Contraseña del usuario","example":"password123"}}},"CreateRecipeDto":{"type":"object","properties":{"title":{"type":"string","description":"Título de la receta","example":"Panqueques"},"description":{"type":"string","description":"Descripción de la receta","example":"Receta para hacer panqueques riquísimos"},"ingredients":{"description":"Lista de ingredientes","example":["Harina","Huevos","Leche"],"type":"array","items":{"type":"string"}},"imageUrl":{"type":"string","description":"URL de la imagen de la receta","example":"https://example.com/panqueques.jpg"},"userId":{"type":"number","description":"ID del usuario que crea la receta","example":1}},"required":["title","description","ingredients","imageUrl","userId"]},"UpdateRecipeDto":{"type":"object","properties":{"title":{"type":"string","description":"Título de la receta","example":"Panqueques"},"description":{"type":"string","description":"Descripción de la receta","example":"Receta para hacer panqueques riquísimos"},"ingredients":{"description":"Lista de ingredientes","example":["Harina","Huevos","Leche"],"type":"array","items":{"type":"string"}},"imageUrl":{"type":"string","description":"URL de la imagen de la receta","example":"https://example.com/panqueques.jpg"}}},"CreateRatingDto":{"type":"object","properties":{"userId":{"type":"number","description":"ID del usuario que hace la valoración","example":1},"recipeId":{"type":"number","description":"ID de la receta que se valora","example":1},"score":{"type":"number","description":"Puntuación de la receta entre 1 y 10","example":8}},"required":["userId","recipeId","score"]},"UpdateRatingDto":{"type":"object","properties":{"score":{"type":"number","description":"Puntuación de la receta entre 1 y 10","example":8}},"required":["score"]}}}}